requirements.txt
roboflow
diffusers==0.23.1
transformers==4.34.0
accelerate==0.23.0
controlnet_aux==0.0.6
torch
torchvision
opencv-python
pycocotools
numpy
matplotlib
clip
ftfy
regex
tqdm
PyYAML
transformers
accelerate
pillow